+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+mail-auth 0.3.4
+================================
+- Added `to_reverse_name` method to `IpAddr` to convert an IP address to a reverse DNS domain name.
+- Added `txt_raw_lookup` method to `Resolver` to perform a raw TXT lookup.
+
 mail-auth 0.3.3
 ================================
 - Bump `mail-parser` dependency to 0.9